product and cart

+++ src/main/java/com/ecommerce/backend/repo/UserRepository.java
@@ -10,6 +10,6 @@ import com.ecommerce.backend.entity.User;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
